MAINT, DEP: Deprecate the `qmflows.settings` module
`Settings` should be directly imported from the main namespace
