+++ src/qmflows/packages/_cp2k.py
@@ -10,7 +10,7 @@ from scm import plams
 
 from ._packages import Package, Result, parse_output_warnings, load_properties
 from ..parsers.cp2k import parse_cp2k_warnings
-from ..settings import Settings
+from .._settings import Settings
 from ..warnings_qmflows import cp2k_warnings, Key_Warning
 from ..type_hints import Final, _Settings, Generic2Special
 from ..common import InfoMO
